# backend/internal/entities/tokens_entity.go
package entities

import (
	"github.com/golang-jwt/jwt/v5"
)

type AccessToken struct {
	Token string `json:"access_token" cookie:"access_token"`
}
type RefreshToken struct {
	Token string `json:"refresh_token" cookie:"refresh_token" redis:"refresh_token"`
}

type TokenPair struct {
	AccessToken  string `json:"access_token" cookie:"access_token"`
	RefreshToken string `json:"refresh_token" cookie:"refresh_token"`
}

type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token" cookie:"refresh_token"`
}

type TokenClaims struct {
	UserID int    `json:"user_id"`
	Type   string `json:"type"`
	jwt.RegisteredClaims
}